Added login feature
In login view whenever you press submit, it sends a post request to '/api/authenticate'. On the serverside it'll use try to match that up with the database and if it matches it'll send the users info back to the frontend. The users info will then get saved to $rootScope.globals